fix noise augmentation

+++ data_utils/ModelNetDataLoader.py
@@ -190,8 +190,8 @@ class ModelNetDataLoader(Dataset):
                 point_set = point_set[0:self.npoints, :]
 
             if random.random() < self.noise_augmentation_stddev:
-                noise = np.random.normal(0,self.noise_augmentation_stddev,point_set.shape)
-                # noise = torch.randn(point_set.shape) * self.noise_augmentation_stddev
+                noise = np.random.normal(0, self.noise_augmentation_stddev, point_set.shape).astype(np.float32)
+                # noise = torch.randn(point_set.shape, dtype=point_set.dtype) * self.noise_augmentation_stddev
                 point_set = point_set + noise
 
             if random.random() < self.rotation_augmentation_probability:
